Treat unused function arguments as error in MSVC
And fix errors caused by this. In a public header, one would want
to have as few warnings as possible, as users of the header
will see the warnings, and some might treat warnings as errors
in their own code.

Also this change might have revealed a bug in UTPP:
ReporterDeferred::SuiteStart() was ignoring the suite
argument, contrary to what the documentation above
suggested.