Adjust formatting of error lines so that VC++ and others show them as errors
UnitTest++ formats its lines similarly.

If running unittests as a post-build action in VC++, properly formatted
lines (i.e. ones that have the word 'error' in them besides file and line
number) are shown as errors in the Errors tab in the IDE. Other kind of
lines are just shown in the Output.